LibWeb: Split ColorStyleValue out of StyleValue.{h,cpp}
